Fix OSGi Bundle deploy by setting symbolicName from manifest
Apply OSGi changes from cics-bundle-common 0.0.3. Name moved higher in binding hierarchy. It took me a while to figure out what was going on with the two Artifact classes, so I tried to make it clearer. test-no-transitive converted to war just so I didn't have to bother with a manifest file.

Signed-off-by: Tom Foyle <tom.foyle@uk.ibm.com>